+++ resources/js/pages/Files/Browser.vue
@@ -4,8 +4,25 @@
             <router-link :to="{ name: 'apps.edit', params: { id: site.id }}"
                 is="sui-button" color="teal" icon="globe" floated="right"
                 id="back2site" :content="'Edit ' + site.name" v-if="site" />
-            <sui-button id="levelup" size="mini" icon="level up" @click="upOneLevel"/>
-            {{ currentPath }}
+
+            <sui-button id="levelup" icon="level up" @click="upOneLevel"/>
+
+            <sui-breadcrumb class="massive">
+                <template v-for="(segment, index) in pathParts">
+
+                    <sui-breadcrumb-section link @click="setPath(segment.path)"
+                        v-if="segment.path != currentPath">
+                        {{ segment.dirname }}
+                    </sui-breadcrumb-section>
+                    <sui-breadcrumb-section v-else>
+                        {{ segment.dirname }}
+                    </sui-breadcrumb-section>
+
+                    <sui-breadcrumb-divider @click="setPath(segment.path)"
+                        v-if="index < (pathParts.length - 1)" />
+
+                </template>
+            </sui-breadcrumb>
         </h2>
 
         <file-list :files="files" @set-path="setPath($event)" />
@@ -21,6 +38,10 @@ export default {
         this.$store.dispatch('loadSites');
         this.$store.dispatch('loadFiles', { path: this.path });
     },
+    beforeRouteUpdate (to, from, next) {
+        this.$store.dispatch('loadFiles', { path: to.params.path });
+        next();
+    },
     props: [
         'path',
     ],
@@ -33,17 +54,35 @@ export default {
             'getSiteByDocroot',
             'files',
         ]),
+        pathParts: function() {
+            let parts = [],
+                path = '';
+
+            for (let part of this.currentPath.split('/')) {
+                path = path + part + '/';
+
+                parts.push({
+                    'path': path.replace(/\/+$/, ''),
+                    'dirname': part,
+                });
+            }
+
+            return parts;
+        },
         site: function() {
             return this.getSiteByDocroot(this.currentPath);
         },
     },
     methods: {
-        setPath: function (file) {
-            let path = this.currentPath == '/' ? '/' + file.filename
-                     : this.currentPath + '/' + file.filename
+        setPath: function (target) {
+            if (typeof target == 'string') {
+                target = '' == target ? '/' : target;
+            } else {
+                target = this.currentPath == '/' ? '/' + target.filename
+                     : this.currentPath + '/' + target.filename
+            }
 
-            this.$router.push({ name: 'files', params: { path: path } });
-            this.$store.dispatch('loadFiles', { path: path });
+            this.$router.push({ name: 'files', params: { path: target } });
         },
         upOneLevel: function () {
             let path = this.currentPath;